# backend/api/openapi.yaml
openapi: 3.0.3
info:
  title: VAN API
  version: 1.0.0
  description: >
    API REST de VAN para la app móvil (Flutter). Toda la lógica de negocio
    vive en el servidor; la app es un cliente delgado.

    Autenticación: Bearer JWT. El login devuelve access_token (corto,
    30 min) y refresh_token (opaco, 30 días, rotativo). Cuando el
    access_token expira, la app debe llamar POST /auth/refresh con el
    refresh_token para obtener un par nuevo. El refresh_token se revoca
    con POST /auth/logout.

    Envelope de respuestas:
      Exito:  {"success": true,  "data": {...}}
      Error:  {"success": false, "error": {"code": "CODE", "message": "..."}}
...
paths:
...
  /trips/{id}:
    get:
      tags: [trips]
      summary: Detalle del viaje (payload canónico). Solo participantes.
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/TripId'
      responses:
        '200':
          description: Viaje
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TripData'
        '403': { $ref: '#/components/responses/Error' }
        '404': { $ref: '#/components/responses/Error' }
    parameters:
      - $ref: '#/components/parameters/TripId'
  /trips/available:
    get:
      tags: [trips]
      summary: Viajes requested cercanos (modo conductor).
      description: |
        lat/lng del query tienen prioridad; si faltan se usa la posición
        persistida del perfil; sin ninguna → LOCATION_REQUIRED. Expira
        primero los requested con más de 5 minutos (cancelled_by=system).
      security:
        - bearerAuth: []
      parameters:
        - {name: lat, in: query, schema: {type: number}}
        - {name: lng, in: query, schema: {type: number}}
        - {name: radius, in: query, schema: {type: number, default: 10}, description: km}
        - {name: vehicle_type, in: query, schema: {type: string, enum: [moto, auto]}}
        - {name: page, in: query, schema: {type: integer, default: 1}}
        - {name: limit, in: query, schema: {type: integer, default: 20, maximum: 100}}
      responses:
        '200':
          description: Items ordenados por distancia al pickup
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: {type: boolean}
                  data:
                    type: object
                    properties:
                      items:
                        type: array
                        items:
                          $ref: '#/components/schemas/AvailableTrip'
                      pagination:
                        $ref: '#/components/schemas/Pagination'
                      driver_lat: {type: number, nullable: true}
                      driver_lng: {type: number, nullable: true}
        '400': { $ref: '#/components/responses/Error' }
  /trips/{id}/accept:
    post:
      tags: [trips]
      summary: Conductor toma el viaje (claim atómico). Requiere online+libre+approved.
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/TripId'
      responses:
        '200':
          description: Aceptado — is_busy=true en el conductor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TripData'
        '403': { $ref: '#/components/responses/Error', description: NOT_VERIFIED }
        '404': { $ref: '#/components/responses/Error' }
        '409': { $ref: '#/components/responses/Error', description: TRIP_NOT_AVAILABLE o NOT_ONLINE }
  /trips/{id}/reject:
    post:
      tags: [trips]
      summary: Rechaza un viaje ofrecido (sin efecto sobre él).
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/TripId'
      responses:
        '200':
          description: '{ok: true}'
          content:
            application/json:
              schema: { $ref: '#/components/schemas/OkResponse' }
        '404': { $ref: '#/components/responses/Error' }
  /trips/{id}/start:
    post:
      tags: [trips]
      summary: Inicia el viaje (accepted → ongoing). Solo conductor asignado.
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/TripId'
      responses:
        '200': 
          description: ongoing con started_at
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TripData'
        '403': { $ref: '#/components/responses/Error' }
        '409': { $ref: '#/components/responses/Error', description: INVALID_TRANSITION }
  /trips/{id}/complete:
    post:
      tags: [trips]
      summary: Cobra y finaliza (ongoing → completed) vía finalize_trip().
      description: |
        Body {} usa el payment_method elegido por el pasajero. El conductor
        puede confirmar el método real con {"method": ...} (misma regla que
        la web). Billetera sin saldo → INSUFFICIENT_BALANCE y el viaje
        sigue ongoing para reintentar con otro método. Retry idempotente.
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/TripId'
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                method:
                  type: string
                  enum: [efectivo, mercadopago, transferencia, tarjeta, billetera]
      responses:
        '200':
          description: completed con fare.final y wallet
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TripData'
        '400': { $ref: '#/components/responses/Error', description: INVALID_PAYMENT_METHOD o INSUFFICIENT_BALANCE }
        '409': { $ref: '#/components/responses/Error', description: INVALID_TRANSITION }
  /trips/{id}/cancel:
    post:
      tags: [trips]
      summary: Cancela (pasajero dueño o conductor asignado). Libera is_busy.
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/TripId'
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                reason: {type: string, maxLength: 500}
      responses:
        '200':
          description: cancelled con cancelled_by=passenger|driver
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TripData'
        '403': { $ref: '#/components/responses/Error' }
        '409': { $ref: '#/components/responses/Error', description: TRIP_FINALIZED }
  /trips/{id}/rate:
    post:
      tags: [trips]
      summary: Califica el viaje completado (1 vez por rol).
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/TripId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [rating]
              properties:
                rating: {type: integer, minimum: 1, maximum: 5}
                comment: {type: string, maxLength: 500}
      responses:
        '200':
          description: '{ok: true}'
          content:
            application/json:
              schema: { $ref: '#/components/schemas/OkResponse' }
        '400': { $ref: '#/components/responses/Error', description: INVALID_RATING }
        '403': { $ref: '#/components/responses/Error' }
        '409': { $ref: '#/components/responses/Error', description: TRIP_NOT_COMPLETED o ALREADY_RATED }
  /trips/{id}/eta:
    get:
      tags: [trips]
      summary: ETA del conductor asignado hacia el pickup.
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/TripId'
      responses:
        '200':
          description: Valores null si aún no hay conductor o posición
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: {type: boolean}
                  data:
                    type: object
                    properties:
                      eta_min: {type: integer, nullable: true}
                      distance_km: {type: number, nullable: true}
                      driver_lat: {type: number, nullable: true}
                      driver_lng: {type: number, nullable: true}
        '403': { $ref: '#/components/responses/Error' }
        '404': { $ref: '#/components/responses/Error' }
...
  /trips/{id}/status:
    get:
      tags: [trips]
      summary: Estado del viaje + ubicación del conductor (polling cada 5s)
      description: Payload ligero optimizado para polling. Solo participantes.
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/TripId'
      responses:
        '200':
          description: Estado del viaje
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/OkResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          id: {type: integer}
                          status: {type: string}
                          vehicle_type: {type: string}
                          pickup_address: {type: string}
                          dropoff_address: {type: string}
                          pickup_lat: {type: number, nullable: true}
                          pickup_lng: {type: number, nullable: true}
                          dropoff_lat: {type: number, nullable: true}
                          dropoff_lng: {type: number, nullable: true}
                          fare: {type: string}
                          distance_km: {type: number, nullable: true}
                          payment_method: {type: string}
                          payment_status: {type: string, nullable: true}
                          payment_method_collected: {type: string, nullable: true}
                          driver:
                            type: object
                            nullable: true
                            properties:
                              id: {type: integer}
                              name: {type: string}
                              phone: {type: string}
                              profile_picture: {type: string, nullable: true}
                              rating_avg: {type: number}
                              rating_count: {type: integer}
                              lat: {type: number, nullable: true}
                              lng: {type: number, nullable: true}
                              vehicle_info:
                                type: object
                                nullable: true
                                properties:
                                  type: {type: string}
                                  placa: {type: string}
                                  marca: {type: string}
                                  modelo: {type: string}
        '403': { $ref: '#/components/responses/Error' }
        '404': { $ref: '#/components/responses/Error' }
  # ─── A3: Auth profile + password + guidelines ───
...
  /trips:
    get:
      tags: [trips]
      summary: Listar viajes propios (paginado)
      security:
        - bearerAuth: []
      parameters:
        - {name: role, in: query, schema: {type: string, enum: [passenger, driver]}}
        - {name: status, in: query, schema: {type: string}}
        - {name: page, in: query, schema: {type: integer, default: 1}}
        - {name: limit, in: query, schema: {type: integer, default: 20}}
      responses:
        '200':
          description: Viajes propios
components:
  parameters:
    TripId:
      name: id
      in: path
      required: true
      schema: {type: integer}
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    Error:
      description: Error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    ErrorResponse:
      type: object
      properties:
        success: {type: boolean, example: false}
        error:
          type: object
          properties:
            code:
              type: string
              example: INVALID_CREDENTIALS
            message:
              type: string
              example: Credenciales inválidas
...
    User:
      type: object
      properties:
        id: {type: integer}
        type: {type: string, enum: [user, driver]}
        name: {type: string}
        email: {type: string}
        phone: {type: string}
        email_verified: {type: boolean}
        rating_avg: {type: number}
        rating_count: {type: integer}
        profile_picture: {type: string, nullable: true}
        vehicle_type: {type: string, nullable: true}
        is_verified: {type: boolean, nullable: true}
...
    FareBreakdown:
      type: object
      description: Dinero como string decimal; platform_fee_rate null = comisión desactivada
      properties:
        total_fare: {type: string, example: '11.25'}
        platform_fee: {type: string, example: '0.56'}
        platform_fee_rate: {type: string, nullable: true, example: '0.05'}
        driver_earnings: {type: string, example: '10.69'}
        currency: {type: string, example: ARS}
    Trip:
      type: object
      properties:
        id: {type: integer}
        status: {type: string, enum: [requested, accepted, ongoing, completed, cancelled]}
        vehicle_type: {type: string, enum: [moto, auto]}
        pickup_address: {type: string}
        dropoff_address: {type: string}
        pickup_lat: {type: number, nullable: true}
        pickup_lng: {type: number, nullable: true}
        dropoff_lat: {type: number, nullable: true}
        dropoff_lng: {type: number, nullable: true}
        distance_km: {type: number, nullable: true}
        duration_min: {type: integer, nullable: true}
        payment_method: {type: string}
        company_id: {type: integer, nullable: true}
        requested_at: {type: string, format: date-time, nullable: true}
        started_at: {type: string, format: date-time, nullable: true}
        completed_at: {type: string, format: date-time, nullable: true}
        cancelled_by: {type: string, nullable: true}
        fare:
          type: object
          properties:
            estimate:
              $ref: '#/components/schemas/FareBreakdown'
            final:
              $ref: '#/components/schemas/FareBreakdown'
              nullable: true
              description: Solo cuando status == completed
        wallet:
          type: object
          description: Etapa 4 — hoy charged es siempre false al crear
          properties:
            charged: {type: boolean}
            passenger_txn_id: {type: integer, nullable: true}
            driver_txn_id: {type: integer, nullable: true}
        passenger:
          $ref: '#/components/schemas/User'
        driver:
          allOf:
            - $ref: '#/components/schemas/User'
          nullable: true
          description: null hasta que un conductor acepte
...
    TripData:
      type: object
      description: Envoltorio estándar con el viaje
      properties:
        success: {type: boolean}
        data:
          type: object
          properties:
            trip:
              $ref: '#/components/schemas/Trip'
    OkResponse:
      type: object
      properties:
        success: {type: boolean}
        data:
          type: object
          properties:
            ok: {type: boolean}
    Pagination:
      type: object
      properties:
        page: {type: integer}
        limit: {type: integer}
        total: {type: integer}
        pages: {type: integer}
    AvailableTrip:
      type: object
      description: Item ligero de GET /trips/available (contrato §5.2.4)
      properties:
        id: {type: integer}
        pickup_address: {type: string}
        dropoff_address: {type: string}
        pickup_lat: {type: number, nullable: true}
        pickup_lng: {type: number, nullable: true}
        vehicle_type: {type: string, enum: [moto, auto]}
        payment_method: {type: string}
        fare:
          type: object
          properties:
            estimate:
              $ref: '#/components/schemas/FareBreakdown'
        distance_km: {type: number, description: Distancia del conductor al pickup}
        requested_at: {type: string, format: date-time}